Add onValueChanged
